formatting

+++ api/routers/users.py
@@ -55,9 +55,7 @@ def create_access_token(data: dict, expires_delta: datetime.timedelta):
     to_encode = data.copy()
     expire = datetime.datetime.now(datetime.UTC) + expires_delta
     to_encode.update({"exp": expire})
-    encoded_jwt = jwt.encode(
-        to_encode, str(TOKEN_SECRET_KEY), algorithm=TOKEN_ALGORITHM or "HS256"
-    )
+    encoded_jwt = jwt.encode(to_encode, str(TOKEN_SECRET_KEY), algorithm=TOKEN_ALGORITHM or "HS256")
     return encoded_jwt
 
 
@@ -221,14 +219,10 @@ async def update_user(user_update: UserUpdate, token: str = Header(None)):
         update_fields["password"] = update_fields["password"]
 
     if "currencies" in update_fields and isinstance(update_fields["currencies"], list):
-        new_currencies = list(
-            set(user.get("currencies", []) + update_fields["currencies"])
-        )
+        new_currencies = list(set(user.get("currencies", []) + update_fields["currencies"]))
         update_fields["currencies"] = new_currencies
     try:
-        result = await users_collection.update_one(
-            {"_id": ObjectId(user_id)}, {"$set": update_fields}
-        )
+        result = await users_collection.update_one({"_id": ObjectId(user_id)}, {"$set": update_fields})
         if result.modified_count == 1:
             updated_user = await users_collection.find_one({"_id": ObjectId(user_id)})
             return {
@@ -321,17 +315,13 @@ async def get_token(token_id: str, token: str = Header(None)) -> dict:
         dict: Details of the specified token.
     """
     user_id = await verify_token(token)
-    token_data = await tokens_collection.find_one(
-        {"user_id": user_id, "_id": ObjectId(token_id)}
-    )
+    token_data = await tokens_collection.find_one({"user_id": user_id, "_id": ObjectId(token_id)})
     if not token_data:
         raise HTTPException(status_code=404, detail="Token not found")
 
     formatted_token = format_id(token_data)
     # Convert datetime to ISO format
-    if "expires_at" in formatted_token and isinstance(
-        formatted_token["expires_at"], datetime.datetime
-    ):
+    if "expires_at" in formatted_token and isinstance(formatted_token["expires_at"], datetime.datetime):
         formatted_token["expires_at"] = formatted_token["expires_at"].isoformat()
 
     return formatted_token
@@ -379,9 +369,7 @@ async def delete_token(token_id: str, token: str = Header(None)):
         dict: Message indicating whether the token was successfully deleted.
     """
     user_id = await verify_token(token)
-    result = await tokens_collection.delete_one(
-        {"user_id": user_id, "_id": ObjectId(token_id)}
-    )
+    result = await tokens_collection.delete_one({"user_id": user_id, "_id": ObjectId(token_id)})
 
     if result.deleted_count == 1:
         return {"message": "Token deleted successfully"}
